requirements.txt
alembic==1.4.3
aniso8601==8.1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Deprecated==1.2.10
Flask==0.12.5
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
PyGithub==1.54.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
urllib3==1.26.2
Werkzeug==0.10.4
wrapt==1.12.1